+++ tesseraql-identity/src/main/resources/io/tesseraql/identity/pack/default/sql/find-roles-by-user-id.sql
@@ -9,8 +9,8 @@ from
   join tql_roles r on r.role_id = ur.role_id
 where
   ur.user_id = /* userId */ 'u1'
-  and (ur.starts_at is null or ur.starts_at <= current_timestamp)
-  and (ur.ends_at is null or ur.ends_at > current_timestamp)
+  and (ur.starts_at is null or ur.starts_at <= /* now */ '2026-08-20 09:00:00')
+  and (ur.ends_at is null or ur.ends_at > /* now */ '2026-08-20 09:00:00')
 
 union
 
@@ -26,6 +26,6 @@ from
   join tql_roles r on r.role_id = gr.role_id
 where
   ug.user_id = /* userId */ 'u1'
-  and (ug.starts_at is null or ug.starts_at <= current_timestamp)
-  and (ug.ends_at is null or ug.ends_at > current_timestamp)
+  and (ug.starts_at is null or ug.starts_at <= /* now */ '2026-08-20 09:00:00')
+  and (ug.ends_at is null or ug.ends_at > /* now */ '2026-08-20 09:00:00')
 ;