+++ tesseraql-identity/src/main/resources/io/tesseraql/identity/pack/default/sql/list-groups.sql
@@ -8,8 +8,8 @@ select
   g.tenant_id  as tenant_id,
   (select count(*) from tql_user_groups ug
      where ug.group_id = g.group_id
-       and (ug.starts_at is null or ug.starts_at <= current_timestamp)
-       and (ug.ends_at is null or ug.ends_at > current_timestamp)) as member_count
+       and (ug.starts_at is null or ug.starts_at <= /* now */ '2026-08-20 09:00:00')
+       and (ug.ends_at is null or ug.ends_at > /* now */ '2026-08-20 09:00:00')) as member_count
 from
   tql_groups g
 order by